moved message logging to its own package
